fix(cuda): grouped MoE prefill — bit-exact output (u64 cache offsets, serial scatter)
Two bugs were keeping the grouped-MoE prefill path opt-in:

1. **u32 cache-offset overflow (the actual blocker).**
   `gate_packed_offsets` etc. were `Vec<u32>` / `DeviceBuffer<u32>`, but the
   VRAM expert cache buffer is ~9 GB on Gemma-4-26B. Past layer ~10 the
   cumulative `packed_offset` exceeded 4 GB and silently truncated to 32 bits,
   so the grouped GEMM read weights from a completely different layer in the
   cache. The earlier per-stage probes only verified the first few layers and
   missed this. Switched the host vectors, device buffers, kernel parameter
   types, and Rust wrappers to `u64` / `unsigned long long *`. Added
   `alloc_u64` / `upload_u64_slice_to_device` to the runtime.

2. **Float reduction-order divergence.**
   The fused `unpermute_scatter_add` kernel relies on atomicAdd ordering across
   blocks, which is non-deterministic and introduces 1-ULP per-element drift —
   Gemma 4's softmax+top-k router amplifies that into flipped expert
   selections within a few layers. Replaced it with a per-expert serial scatter
   that copies each cached expert's `permuted_down` slice into `gather_out` and
   reuses the existing `scatter_add_weighted_f32_device`, iterating
   `e = 0..num_experts` to match the per-expert fallback's accumulation order
   exactly. Prefix sum of `cached_counts` is done on the host from
   `counts_host` so there's no extra GPU→host sync per layer.

   Likewise, `aegis_nvfp4_linear_prequantized_batched` (the single-batch kernel
   used for `count = 1`) is no longer dispatched — the gemm-style kernel now
   handles every batch size. The per-expert and grouped paths now share an
   identical `aegis_nvfp4_prequant_gemm_inner` `__forceinline__` device
   function for the dot product, so both kernels compile to identical SASS.

Result: `quality-diff` PASSes for both default and `AEGIS_GROUPED_MOE_ENABLE=1`.
Gemma 4 generates "I am Gemma 4, a large language model developed by Google
DeepMind…" identically in both modes.

Performance on Gemma-4-26B-A4B-NVFP4, RTX 5070 Ti, 720-token prompt:

| config                                    | prefill | decode |
|-------------------------------------------|---------|--------|
| cache off                                 | 112.6   | 41.8   |
| cache on, fallback (default)              | 113.4   | 43.9   |
| cache on, grouped (AEGIS_GROUPED_MOE_ENABLE=1) | 110.2   | 43.9   |

Grouped is ~3% slower than fallback at this prompt size — the matmul inner
loop is still 1-warp-per-output-row scalar fp32, so wall time is dominated by
NVFP4 decode + fma rather than launch overhead. The grouped dispatcher is now
correct and ready for a tensor-core (mma.sync / CUTLASS NVFP4) inner kernel
to replace `aegis_nvfp4_prequant_gemm_inner` — that is the real path to
"thousands of tps".

+++ crates/aegisllm-cuda/src/cuda/runtime/ops.rs
@@ -413,9 +413,9 @@ impl CudaRuntime {
     pub fn nvfp4_grouped_prequant_gemm_device(
         &self,
         base_packed: &DeviceBuffer<u8>,
-        packed_offsets: &DeviceBuffer<u32>,
+        packed_offsets: &DeviceBuffer<u64>,
         base_scales: &DeviceBuffer<u8>,
-        scales_offsets: &DeviceBuffer<u32>,
+        scales_offsets: &DeviceBuffer<u64>,
         expert_counts: &DeviceBuffer<u32>,
         expert_first_token_off: &DeviceBuffer<u32>,
         expert_output_scales: &DeviceBuffer<f32>,
